kernel-3.x.x: tun: Fixed unsigned/signed comparison, commit 87897931 upstream

+++ trunk/linux-3.4.x/drivers/net/tun.c
@@ -979,7 +979,7 @@ static int tun_recvmsg(struct kiocb *iocb, struct socket *sock,
 		return -EINVAL;
 	ret = tun_do_read(tun, iocb, m->msg_iov, total_len,
 			  flags & MSG_DONTWAIT);
-	if (ret > total_len) {
+	if (ret > (ssize_t)total_len) {
 		m->msg_flags |= MSG_TRUNC;
 		ret = flags & MSG_TRUNC ? ret : total_len;
 	}
